+++ RELEASE_NOTES.md
@@ -1,3 +1,6 @@
+#### Unreleased ####
+* Added `Offset.FromEnd` support to current and live events-by-tag and all-events queries.
+
 #### 1.5.70 July 6th, 2026 ####
 * [Bumped Akka to version 1.5.70](https://github.com/akkadotnet/Akka.Persistence.EventStore/pull/88)
 * [Fixed an issue with concurrent writes for persistent subscriptions](https://github.com/akkadotnet/Akka.Persistence.EventStore/pull/83)